+++ src/main/kotlin/com/github/filiplabs/devpulse/toolWindow/DevPulseToolWindowFactory.kt
@@ -9,8 +9,8 @@
 
 package com.github.filiplabs.devpulse.toolWindow
 
-import com.intellij.openapi.util.Disposer
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.Disposer
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
 import com.intellij.ui.content.ContentFactory