+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ## [Unreleased]
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- DevPulse focus tracking, edit breakdown, top files, and Pomodoro dashboard.